silence shellcheck + yamllint
Actually reworking the sh code would be an alternative, but this commit
at least checks in a linter config corresponding to the coding style
used.

No yaml files, apart from the linter config itself, are checked in to
this repo, but running the tests generates one with overly long lines.